refactor(globals): one PascalCase folder per global
Dissolve the sahaj-atlas/, wemeditate-app/, wemeditate-web/ project
groups into 7 per-global folders named after their TS export. The
WeMeditate App status global becomes WeMeditateAppStatus/ with its
section builders under sections/ and statusConfig.json co-located.

Cross-folder constants (VIBE_CHECK_IDENTIFIERS, APP_REQUIRED_PAGE_FIELDS,
translationsSchema.json) are imported via @/globals/<Name> per the
straight-move decision. Update the master barrel, the Pages collection,
generate-status-config.ts, the translation seeds, and the global tests.

No per-global index.ts files; no behavior change.

Refs #442
