+++ seeds/translations/import.ts
@@ -16,9 +16,9 @@
 
 import * as path from 'path'
 
-import atlasSchema from '../../src/globals/sahaj-atlas/translationsSchema.json' with { type: 'json' }
-import appSchema from '../../src/globals/wemeditate-app/translationsSchema.json' with { type: 'json' }
-import wmWebSchema from '../../src/globals/wemeditate-web/translationsSchema.json' with { type: 'json' }
+import atlasSchema from '../../src/globals/SahajAtlasTranslations/translationsSchema.json' with { type: 'json' }
+import appSchema from '../../src/globals/WeMeditateAppTranslations/translationsSchema.json' with { type: 'json' }
+import wmWebSchema from '../../src/globals/WeMeditateWebTranslations/translationsSchema.json' with { type: 'json' }
 import { BaseImporter, type BaseImportOptions } from '../lib'
 import {
   buildWmAppGlobalData,
@@ -26,7 +26,6 @@ import {
   type SeedFile,
   type TranslationsSchemaRoot,
 } from '../wm-app-translations/lexicalConverter'
-
 
 // ============================================================================
 // Example-data generator (for wm-web and sy-atlas)
