+++ src/collections/content/Pages.ts
@@ -2,7 +2,7 @@ import type { CollectionConfig } from 'payload'
 
 import { pageBlocks } from '@/blocks'
 import { slugField } from '@/fields'
-import { APP_REQUIRED_PAGE_FIELDS } from '@/globals/wemeditate-app/config'
+import { APP_REQUIRED_PAGE_FIELDS } from '@/globals/WeMeditateAppConfig/WeMeditateAppConfig'
 import { removeDanglingLexicalReferencesAfterRead } from '@/hooks/lexicalHooks'
 import { PAGE_TAGS } from '@/lib/constants'
 import { fullRichTextEditor } from '@/lib/richEditor'
