+++ src/globals/WeMeditateAppConfig/WeMeditateAppConfig.ts
