+++ src/globals/WeMeditateAppStatus/sections/appCards.ts
