+++ src/globals/WeMeditateAppStatus/sections/translations.ts
@@ -1,7 +1,7 @@
+import translationsSchema from '@/globals/WeMeditateAppTranslations/translationsSchema.json' with { type: 'json' }
 import { getLocaleLabel, isValidLocale } from '@/lib/locales'
 import { type GroupSpec, type SectionSpec } from '@/lib/status'
 
-import translationsSchema from '../translationsSchema.json' with { type: 'json' }
 import { type WeMeditateAppStatusConfig } from './shared'
 
 type LeafProp = { type: 'string' | 'richText' }
