+++ src/globals/WeMeditateAppStatus/statusConfig.json
