+++ src/globals/WeMeditateAppTranslations/translationsSchema.json
