+++ src/globals/WeMeditateWebTranslations/translationsSchema.json
